I guess fixed print fps on delay
Bit wack, we can assign a time thread

+++ Engine/source/engine/components/time/Time.h
@@ -22,7 +22,7 @@ namespace Engine {
 		*/
 		static void UpdateTime();
 		static int GetFramesPerSecond();
-		static int GetFramesPerSecond(float delay);
+		static float  GetFramesPerSecond(float delay);
 
 	private:
 		static void UpdateDeltaTime();
